Big refactor & work
